+++ grpcalchemy/__init__.py
@@ -10,5 +10,5 @@ __email__ = 'guangtian_li@qq.com'
 __version__ = '0.2.4'
 
 from .server import Server
-from .blueprint import Blueprint, Context
+from .blueprint import Blueprint, Context, current_rpc
 from .globals import current_app